Fix tuple unpacking matches
PR #99 attempted to fix the handling of multiple definitions by way of
tuple unpackings -- eg
(x,y,z) = ... and
(x,
 y) = undefined

Unfortunately, this was an edge case underexercised in the test suite,
so the bugs in #99 weren't caught -- we were lucky it also broke the
handling of operators, so 9.hs caught it.

Moreover, the comma splitting code was impenetrably obscure, rewrote it
as a simple unfold.

Closes: #101
Fixes: #99